@@ -6,13 +6,14 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 public class CommandColor implements TabExecutor {
 
     @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+    public boolean onCommand(CommandSender sender, @NotNull Command command, @NotNull String label, String[] args) {
         if (!sender.hasPermission("servercore.donator")) {
             sender.sendMessage(ChatColor.RED + "You need to donate to change your name color!");
             return true;
@@ -31,7 +32,6 @@ public class CommandColor implements TabExecutor {
             sender.sendMessage(
                     ChatColor.RED + "That is not a valid color option. See /colors for valid color IDs to use.");
         }
-
         return true;
     }
 
