@@ -10,17 +10,12 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
+import org.jetbrains.annotations.NotNull;
 
-public class CommandDupe implements CommandExecutor {
-
-    private final ServerCore serverCore;
-
-    public CommandDupe(ServerCore serverCore) {
-        this.serverCore = serverCore;
-    }
+public record CommandDupe(ServerCore serverCore) implements CommandExecutor {
 
     @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+    public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, String[] args) {
         if (!serverCore.config.getBoolean("enable-dupe")) {
             sender.sendMessage(ChatColor.RED + "This is currently disabled. Check back later.");
             return true;
@@ -29,7 +24,7 @@ public class CommandDupe implements CommandExecutor {
         if (args.length == 0) {
             sender.sendMessage(ChatColor.GREEN + "You have "
                     + serverCore.dupeChargesFUCK.getDupeCharges(
-                            Bukkit.getPlayerExact(sender.getName()).getUniqueId()) + " dupe charges left.");
+                    Bukkit.getPlayerExact(sender.getName()).getUniqueId()) + " dupe charges left.");
             sender.sendMessage(
                     ChatColor.GREEN + "Type /dupe confirm to confirm you want to dupe what you are holding.");
             return true;
