@@ -5,11 +5,12 @@ import net.md_5.bungee.api.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.jetbrains.annotations.NotNull;
 
 public class CommandKillWitherSkulls implements CommandExecutor {
 
     @Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+    public boolean onCommand(CommandSender sender, @NotNull Command command, @NotNull String label, String[] args) {
         if (sender.isOp()) {
             FuckWitherSkulls.killWitherSkulls();
             sender.sendMessage(ChatColor.GOLD + "Killing " + FuckWitherSkulls.countSkulls() + " wither skulls.");
